Refactor: Rename sandbox_flask_app to routes; Clean-up
