--- app/templates/split_result.html
@@ -1,37 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Debt Settlement</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css">
-</head>
-<body>
-    <div class="container mt-5">
-        <h2 class="mb-4">Who Owes Whom – {{ tripname }}</h2>
-
-        {% if result %}
-            <table class="table table-striped">
-                <thead>
-                    <tr>
-                        <th>Debtor</th>
-                        <th>Creditor</th>
-                        <th>Amount (€)</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {% for debtor, creditor, amount in result %}                
-                    <tr>
-                        <td>{{ debtor|title }}</td>
-                        <td>{{ creditor|title }}</td>
-                        <td>{{ '%.2f'|format(amount) }}</td>
-                    </tr>
-                    {% endfor %}
-                </tbody>
-            </table>
-        {% else %}
-            <div class="alert alert-info">No debts to settle – everyone's even!</div>
-        {% endif %}
-
-        <a href="/transactions" class="btn btn-secondary mt-3">← Back to Transactions</a>
-    </div>
-</body>
-</html>
