+++ tests/test_routes.py
@@ -1,6 +1,6 @@
 import pytest
 from pytest_mock import MockFixture
-from app.sandbox_flask_app import app, Group
+from app.routes import app, Group
 from app.database import SessionLocal
 import bcrypt
